VTCCA-904 change organisation id to long (#286)
* make organisationId of type Long

* make organisationId of type Long

+++ app/actions/AuthenticatedAction.scala
@@ -94,7 +94,7 @@ sealed trait AuthenticatedRequest[A] extends Request[A] {
   val organisationAccount: GroupAccount
   val individualAccount: DetailedIndividualAccount
 
-  def organisationId: Int = organisationAccount.id
+  def organisationId: Long = organisationAccount.id
   def personId: Int = individualAccount.individualId
 }
